@@ -1,5 +1,6 @@
-const _ = require('lodash');
-const Type = require('./Type');
+import _ from 'lodash';
+
+import Type from './Type.js';
 
 /**
  * Type Number
@@ -48,4 +49,4 @@ function TypeNumber(validations, typeProvider, validationProvider, messageProvid
   });
 }
 
-module.exports = TypeNumber;
+export default TypeNumber;
