@@ -1,5 +1,6 @@
-const _ = require('lodash');
-const Type = require('./Type');
+import _ from 'lodash';
+
+import Type from './Type.js';
 
 /**
  * Type string
@@ -48,4 +49,4 @@ function TypeString(validations, typeProvider, validationProvider, messageProvid
   });
 }
 
-module.exports = TypeString;
+export default TypeString;
